Messy Jupyter notebook for fitting as many pieces as possible on a given rectangle.
Semi-manual optimization for translation and rotation of a large patch of puzzle pieces to maximize the number of pieces lying completely inside a rectangle of given size.

For an A4 page with 3 mm margins, >10M random samples found 85 placements that fit 56 pieces and none that fit more. Some were manually processed further to choose a point in the middle of the optimal parameter region, and estimate the size of the region to prefer larger ones, using plots of 2D slices of the 3D parameter space. Also in hopes of finding an even better region nearby, but none was found. Rotations cluster around a set of angles 60 degrees apart. No pattern apparent in the translations (apart from obviously the shape of the patch).

Also insert one of the optimal transforms into the SVG (will be overwritten on regeneration).